requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.2
sentence-transformers==2.2.2
pdfplumber==0.10.0
python-docx==1.1.0
groq==0.4.1
pandas>=1.5.0
numpy>=1.21.0
python-multipart==0.0.6
python-dotenv==1.0.0
PyPDF2==3.0.1
plotly==5.17.0
requests==2.31.0
chromadb>=0.4.0
